v1.3.0 + streaming notifications
I added youtube & twitch live notifications into the bot.

It checks the live status for either service once per minute and only makes API requests if it determines the video URL isn't present within the respective live_status.json file.

This reduces API requests to the server and is required for youtube due to the 1k daily limit for free users.

Lastly, I added the googleapis dependency and updated the rest as needed.